task 4, 5, README

+++ ProcessManagement_Lab/processManagement_lab.c
@@ -1,103 +1,31 @@
 #include <strings.h>
-#include <pthread.h>
 #include "processManagement_lab.h"
 
 /**
  * The task function to simulate "work" for each worker process
  * TODO#3: Modify the function to be multiprocess-safe 
  * */
-int code;
-long counter;
 
 void task(long duration)
 {
-    counter = duration;
-    pthread_t thread1 = NULL, thread2 = NULL, thread3 = NULL;
-
     // simulate computation for x number of seconds
     usleep(duration*TIME_MULTIPLIER);
 
+    sem_wait(sem_global_data); // wait
     // TODO: protect the access of shared variable below
     // update global variables to simulate statistics
     ShmPTR_global_data->sum_work += duration;
     ShmPTR_global_data->total_tasks ++;
     if (duration % 2 == 1) {
-        createThread(1, &thread1);
+        ShmPTR_global_data->odd++;
     }
     if (duration < ShmPTR_global_data->min) {
-        createThread(2, &thread2);
+        ShmPTR_global_data->min = duration;
     }
     if (duration > ShmPTR_global_data->max) {
-        createThread(3, &thread3);
+        ShmPTR_global_data->max = duration;
     }
-
-    if (thread1 != NULL)
-    {
-        pthread_join(thread1, NULL);
-    }
-    if (thread2 != NULL)
-    {
-        pthread_join(thread2, NULL);
-    }
-    if (thread3 != NULL)
-    {
-        pthread_join(thread3, NULL);
-    }
-}
-
-void createThread(int threadNum, pthread_t *pThread) {
-    switch (threadNum) {
-        case 1:
-            code = pthread_create(pThread, NULL, &oddDuration, NULL);
-            if (code)
-            {
-                printf("Thread 1 creation failed\n");
-                exit(EXIT_FAILURE);
-            }
-            break;
-        case 2:
-            code = pthread_create(pThread, NULL, &minDuration, NULL);
-            if (code)
-            {
-                printf("Thread 2 creation failed\n");
-                exit(EXIT_FAILURE);
-            }
-            break;
-        case 3:
-            code = pthread_create(pThread, NULL, &maxDuration, NULL);
-            if (code)
-            {
-                printf("Thread 3 creation failed\n");
-                exit(EXIT_FAILURE);
-            }
-            break;
-        default:
-            break;
-    }
-}
-
-void *oddDuration()
-{
-    pthread_mutex_lock(&mutex);
-    ShmPTR_global_data->odd++;
-    pthread_mutex_unlock(&mutex);
-    return NULL;
-}
-
-void *minDuration()
-{
-    pthread_mutex_lock(&mutex);
-    ShmPTR_global_data->min = counter;
-    pthread_mutex_unlock(&mutex);
-    return NULL;
-}
-
-void *maxDuration()
-{
-    pthread_mutex_lock(&mutex);
-    ShmPTR_global_data->max = counter;
-    pthread_mutex_unlock(&mutex);
-    return NULL;
+    sem_post(sem_global_data); // signal
 }
 
 /**
@@ -109,39 +37,31 @@ void job_dispatch(int i){
     //          b. Use semaphore so that you don't busy wait
     //          c. If there's new job, execute the job accordingly: either by calling task(), usleep, exit(3) or kill(getpid(), SIGKILL)
     //          d. Loop back to check for new job
-    int count = 0;
-    while (count < sizeof(shmPTR_jobs_buffer))
+    for(;;)
     {
-        count = 0;
-        for (int i = 0; i < sizeof(shmPTR_jobs_buffer); i++) {
-            switch (shmPTR_jobs_buffer[i].task_status) {
-                case 0:
-                    count++;
+        sem_wait(sem_jobs_buffer[i]);
+        if (shmPTR_jobs_buffer[i].task_status != -1) // not termination
+        {
+            shmPTR_jobs_buffer[i].task_status = 0;
+            switch (shmPTR_jobs_buffer[i].task_type) {
+                case 't':
+                    task(shmPTR_jobs_buffer[i].task_duration);
                     break;
-                case 1:
-                    switch (shmPTR_jobs_buffer[i].task_type) {
-                        case 't':
-                            task(shmPTR_jobs_buffer[i].task_duration);
-                            break;
-                        case 'w':
-                            usleep(shmPTR_jobs_buffer[i].task_duration * TIME_MULTIPLIER);
-                            break;
-                        case 'z':
-                            exit(3);
-                        case 'i':
-                            kill(getpid(), SIGKILL);
-                            break;
-                    }
+                case 'w':
+                    usleep(shmPTR_jobs_buffer[i].task_duration * TIME_MULTIPLIER);
+                    break;
+                case 'z':
                     shmPTR_jobs_buffer[i].task_status = 0;
-                    break;
-                case -1:
+                    exit(3);
+                case 'i':
+                    kill(getpid(), SIGKILL);
                     break;
             }
         }
     }
 
-    printf("Hello from child %d with pid %d and parent id %d\n", i, getpid(), getppid());
-    exit(EXIT_SUCCESS);
+//    printf("Hello from child %d with pid %d and parent id %d\n", i, getpid(), getppid());
+//    exit(EXIT_SUCCESS);
 
 }
 
@@ -230,7 +150,6 @@ void createchildren(){
     // TODO#2:  a. Create number_of_processes children processes
     //          d. For the parent process, continue creating the next children
     //          e. After number_of_processes children are created, return to main
-    pid_t pid;
 
     for (int i = 0; i < number_of_processes; ++i) {
         switch (pid = fork()) {
@@ -238,14 +157,13 @@ void createchildren(){
                 fprintf(stderr, "Fork has failed. Exiting now");
                 exit(EXIT_FAILURE); // exit error
             case 0:
-                //          b. Store the pid_t of children i at children_processes[i]
-                children_processes[i] = pid;
                 //          c. For child process, invoke the method job_dispatch(i)
                 job_dispatch(i);
                 printf("Hello from child %d with pid %d and parent id %d", i, pid, children_processes[0]);
                 exit(EXIT_SUCCESS);
             default:
-                children_processes[0] = pid;
+                //          b. Store the pid_t of children i at children_processes[i]
+                children_processes[i] = pid;
         }
     }
 
@@ -260,26 +178,63 @@ void main_loop(char* fileName){
     // load jobs and add them to the shared memory
     FILE* opened_file = fopen(fileName, "r");
     char action; //stores whether its a 'p' or 'w'
-    long num; //stores the argument of the job 
+    long num; //stores the argument of the job
+    int status;
+    bool assigned = false;
 
     while (fscanf(opened_file, "%c %ld\n", &action, &num) == 2) { //while the file still has input
+        assigned = false;
 
-        //TODO#4: create job, busy wait
-        //      a. Busy wait and examine each shmPTR_jobs_buffer[i] for jobs that are done by checking that shmPTR_jobs_buffer[i].task_status == 0. You also need to ensure that the process i IS alive using waitpid(children_processes[i], NULL, WNOHANG). This WNOHANG option will not cause main process to block when the child is still alive. waitpid will return 0 if the child is still alive. 
-        //      b. If both conditions in (a) is satisfied update the contents of shmPTR_jobs_buffer[i], and increase the semaphore using sem_post(sem_jobs_buffer[i])
-        //      c. Break of busy wait loop, advance to the next task on file 
-        //      d. Otherwise if process i is prematurely terminated, revive it. You are free to design any mechanism you want. The easiest way is to always spawn a new process using fork(), direct the children to job_dispatch(i) function. Then, update the shmPTR_jobs_buffer[i] for this process. Afterwards, don't forget to do sem_post as well 
-        //      e. The outermost while loop will keep doing this until there's no more content in the input file. 
+        while (!assigned)
+        {
+            //TODO#4: create job, busy wait
+            //      a. Busy wait and examine each shmPTR_jobs_buffer[i] for jobs that are done by checking that
+            //      shmPTR_jobs_buffer[i].task_status == 0. You also need to ensure that the process i IS alive using
+            //      waitpid(children_processes[i], NULL, WNOHANG). This WNOHANG option will not cause main process to block
+            //      when the child is still alive. waitpid will return 0 if the child is still alive.
 
+            for (int i = 0; i < number_of_processes; i++)
+            {
+                int alive = waitpid(children_processes[i], &status, WNOHANG);
+                if (shmPTR_jobs_buffer[i].task_status == 0 && alive == 0)
+                {
+                    //      b. If both conditions in (a) is satisfied update the contents of shmPTR_jobs_buffer[i], and
+                    //      increase the semaphore using sem_post(sem_jobs_buffer[i])
+                    assigned = true;
+                    update(i, action, num);
+                    printf("Child process %d with pid %d doing job: %c%ld\n", i, pid, action, num);
+                    //      c. Break of busy wait loop, advance to the next task on file
+                    break;
+                }
+                if (alive != 0) {
+                    //      d. Otherwise if process i is prematurely terminated, revive it. You are free to design any mechanism
+                    //      you want. The easiest way is to always spawn a new process using fork(), direct the children to job_
+                    //      dispatch(i) function. Then, update the shmPTR_jobs_buffer[i] for this process. Afterwards, don't forget
+                    //      to do sem_post as well
+                    printf("child exits prematurely");
+                    createchildren();
+                }
 
+                if (WIFEXITED(status)) {
+                    int es = WEXITSTATUS(status);
 
+                    printf("Exit status was %d\n", es);
+                }
+            }
+            //      e. The outermost while loop will keep doing this until there's no more content in the input file.
+        }
     }
     fclose(opened_file);
 
-    printf("Main process is going to send termination signals\n");
+//    printf("Main process is going to send termination signals\n");
 
     // TODO#4: Design a way to send termination jobs to ALL worker that are currently alive 
-
+    for (int i = 0; i < number_of_processes; i++) {
+        if (waitpid(children_processes[i], &status, WNOHANG) == 0)
+        {
+            update(i, 'z', 0);
+        }
+    }
 
 
     //wait for all children processes to properly execute the 'z' termination jobs
@@ -293,21 +248,66 @@ void main_loop(char* fileName){
     printf("Final results: sum -- %ld, odd -- %ld, min -- %ld, max -- %ld, total task -- %ld\n", ShmPTR_global_data->sum_work, ShmPTR_global_data->odd, ShmPTR_global_data->min, ShmPTR_global_data->max, ShmPTR_global_data->total_tasks);
 }
 
+void update(int count, char action, long num)
+{
+    shmPTR_jobs_buffer[count].task_type = action;
+    shmPTR_jobs_buffer[count].task_duration = num;
+    shmPTR_jobs_buffer[count].task_status = 1;
+    sem_post(sem_jobs_buffer[count]);
+}
+
 void cleanup(){
     //TODO#4: 
     // 1. Detach both shared memory (global_data and jobs)
     // 2. Delete both shared memory (global_data and jobs)
     // 3. Unlink all semaphores in sem_jobs_buffer
+    check_detach(shmdt(ShmPTR_global_data));
+    check_detach(shmdt(shmPTR_jobs_buffer));
+
+    check_rm(shmctl(ShmID_global_data, IPC_RMID, NULL));
+    check_rm(shmctl(ShmID_jobs, IPC_RMID, NULL));
+
+    sem_unlink("semglobaldata");
+    char semjobsi[] = "semjobsi";
+    for (int i = 0; i < number_of_processes; i++) {
+        semjobsi[7] = i + '0';
+        sem_unlink(semjobsi);
+    }
+}
+
+void check_detach(int state)
+{
+    if (state == -1)
+    {
+        printf("Error detaching shared memory.\n");
+    }
+}
+
+void check_rm(int state)
+{
+    if (state == -1)
+    {
+        printf("Error removing shared memory.\n");
+    }
 }
 
 // Real main
 int main(int argc, char* argv[]){
+
+    printf("Lab 1 Starts...\n");
+
+    struct timeval start, end;
+    long secs_used,micros_used;
+
+    //start timer
+    gettimeofday(&start, NULL);
 
     //Check and parse command line options to be in the right format
     if (argc < 2) {
         printf("Usage: sum <infile> <numprocs>\n");
         exit(EXIT_FAILURE);
     }
+
 
     //Limit number_of_processes into 10.
     //If there's no third argument, set the default number_of_processes into 1.
@@ -319,79 +319,21 @@ int main(int argc, char* argv[]){
         else number_of_processes = MAX_PROCESS;
     }
 
-    printf("Number of processes: %d\n", number_of_processes);
-    printf("Main process pid %d \n", getpid());
-
     setup();
     createchildren();
+    main_loop(argv[1]);
 
-    //enter exactly N jobs to the buffer
-    for (int i = 0; i<number_of_processes; i++){
-        shmPTR_jobs_buffer[i].task_type = 't';
-        shmPTR_jobs_buffer[i].task_duration = 1;
-        shmPTR_jobs_buffer[i].task_status = 1; //new, undone job
-        sem_post(sem_jobs_buffer[i]); // signal the child
-    }
+    //parent cleanup
+    cleanup();
 
+    //stop timer
+    gettimeofday(&end, NULL);
 
-    //sleep for 3 seconds, the children processes should all finish by now
-    sleep(3);
+    double start_usec = (double) start.tv_sec * 1000000 + (double) start.tv_usec;
+    double end_usec =  (double) end.tv_sec * 1000000 + (double) end.tv_usec;
 
-    //enter exactly N termination jobs to the buffer
-    for (int i = 0; i<number_of_processes; i++){
-        shmPTR_jobs_buffer[i].task_type = 'z';//termination job
-        shmPTR_jobs_buffer[i].task_duration = 1;
-        shmPTR_jobs_buffer[i].task_status = 1;
-        sem_post(sem_jobs_buffer[i]); // signal the child
-    }
-
-    //wait for all N children processes
-    int waitpid_result;
-    for (int i = 0; i<number_of_processes; i++){
-        waitpid_result = waitpid(children_processes[i], NULL, 0); // returns when child exits normally
-        if (waitpid_result != -1){
-            printf("Child %d with pid %d has exited successfully\n", i, waitpid_result);
-        }
-    }
-
-    // print final results
-    printf("Final results: sum -- %ld, odd -- %ld, min -- %ld, max -- %ld, total task -- %ld\n", ShmPTR_global_data->sum_work, ShmPTR_global_data->odd, ShmPTR_global_data->min, ShmPTR_global_data->max, ShmPTR_global_data->total_tasks);
-
-    //detach and remove shared memory locations
-    int detach_status = shmdt((void *) ShmPTR_global_data); //detach
-    if (detach_status == -1) printf("Detach shared memory global_data ERROR\n");
-    int remove_status = shmctl(ShmID_global_data, IPC_RMID, NULL); //delete
-    if (remove_status == -1) printf("Remove shared memory global_data ERROR\n");
-    detach_status = shmdt((void *) shmPTR_jobs_buffer); //detach
-    if (detach_status == -1) printf("Detach shared memory jobs ERROR\n");
-    remove_status = shmctl(ShmID_jobs, IPC_RMID, NULL); //delete
-    if (remove_status == -1) printf("Remove shared memory jobs ERROR\n");
+    printf("Your computation has used: %lf secs \n", (end_usec - start_usec)/(double)1000000);
 
 
-    //unlink all semaphores before exiting process
-    int sem_close_status = sem_unlink("semglobaldata");
-    if (sem_close_status == 0){
-        printf("Semaphore globaldata closes succesfully.\n");
-    }
-    else{
-        printf("Semaphore globaldata fails to close.\n");
-    }
-
-    for (int i = 0; i<number_of_processes; i++){
-        char *sem_name = malloc(sizeof(char)*16);
-        sprintf(sem_name, "semjobs%d", i);
-        sem_close_status = sem_unlink(sem_name);
-        if (sem_close_status == 0){
-            printf("Semaphore jobs %d closes succesfully.\n", i);
-        }
-        else{
-            printf("Semaphore jobs %d fails to close.\n", i);
-        }
-        free(sem_name);
-    }
-
-
-    printf("success\n");
-
-    return 0;
+    return (EXIT_SUCCESS);
 }
